Add partners and modify swagger

+++ DTOs/TestDto.cs
@@ -2,8 +2,19 @@
 {
     public class TestDto
     {
-        public int Count { get; set; }
+        /// <summary>
+        /// количество
+        /// </summary>
+        public double Count { get; set; }
+
+        /// <summary>
+        /// название
+        /// </summary>
         public string Name { get; set; }
+
+        /// <summary>
+        /// Дата
+        /// </summary>
         public DateTime Date { get; set; }
     }
 }
